UT: Move mongo functions into a module
